kuasa-backend-dj/requirements.txt
apturl==0.5.2
asgiref==3.7.2
azure-core==1.29.4
azure-storage-blob==12.18.2
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.4
django-cors-headers==4.2.0
django-storages==1.14
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
duplicity==0.8.21
fasteners==0.14.1
filelock==3.12.3
future==0.18.2
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
isodate==0.6.1
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pipenv==2023.9.1
platformdirs==3.10.0
protobuf==3.12.4
psycopg2==2.9.7
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rpds-py==0.10.3
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
systemd-python==234
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
uWSGI==2.0.22
vboxapi==1.0
virtualenv==20.24.4
wadllib==1.3.6
xdg==5
xkit==0.0.0
zipp==1.0.0
